+++ Backend/SimulationEngine/UIModel/Components/RenderableComponent.cs
@@ -1,7 +1,7 @@
 ﻿using IdelPog.Common.Structures;
 using IdelPog.ECS.Component;
 
-namespace IdelPog.SimulationEngine.UIModel
+namespace IdelPog.SimulationEngine.UIModel.Components
 {
     public readonly record struct RenderableComponent : IComponent<RenderableComponent>
     {
