+++ Backend/SimulationEngine/UIModel/Service/Screen/IScreenService.cs
@@ -1,6 +1,6 @@
 ﻿using IdelPog.SimulationEngine.UIModel.Structures.Enums;
 
-namespace IdelPog.SimulationEngine.UIModel.Service
+namespace IdelPog.SimulationEngine.UIModel.Service.Screen
 {
     public interface IScreenService
     {
