+++ Frontend/Frontend/MonoGame/RenderingPipeline/Listeners/RenderableDTOListener.cs
@@ -2,6 +2,7 @@
 using System.Collections.Generic;
 using IdelPog.Common.DTO;
 using IdelPog.Frontend.MonoGame.Controllers;
+using IdelPog.Messaging.Listeners;
 using IdelPog.Messaging.Messaging;
 
 namespace IdelPog.Frontend.MonoGame.Listeners
