+++ Shared/Messaging/Assertions/AssertListenerFound.cs
@@ -1,4 +1,5 @@
 ﻿using IdelPog.Messaging.Exceptions;
+using IdelPog.Messaging.Listeners;
 using IdelPog.Messaging.Messaging;
 using IdelPog.Validation.Assertions;
 using IdelPog.Validation.Assertions.Handlers.Interfaces;
