+++ Shared/Messaging/Assertions/Interfaces/IAssertListenerFound.cs
@@ -1,4 +1,5 @@
-﻿using IdelPog.Messaging.Messaging;
+﻿using IdelPog.Messaging.Listeners;
+using IdelPog.Messaging.Messaging;
 
 namespace IdelPog.Messaging.Assertions
 {
