+++ Shared/Messaging/Exceptions/NoListenerFoundException.cs
@@ -1,4 +1,5 @@
-﻿using IdelPog.Messaging.Messaging;
+﻿using IdelPog.Messaging.Listeners;
+using IdelPog.Messaging.Messaging;
 using IdelPog.Validation.Constants;
 
 namespace IdelPog.Messaging.Exceptions
