+++ Shared/Messaging/Messaging/BufferMessenger.cs
@@ -1,4 +1,5 @@
 ﻿using IdelPog.Messaging.Assertions;
+using IdelPog.Messaging.Listeners;
 using IdelPog.Validation.Assertions.Interfaces;
 
 namespace IdelPog.Messaging.Messaging
