+++ Shared/Messaging/Messaging/IBufferMessenger.cs
@@ -1,4 +1,6 @@
-﻿namespace IdelPog.Messaging.Messaging
+﻿using IdelPog.Messaging.Listeners;
+
+namespace IdelPog.Messaging.Messaging
 {
     public interface IBufferMessenger
     {
